refactor(clerk-js): Extract and reuse PreviewButton in UserButton & OrgSwitcher
