+++ packages/clerk-js/src/ui/components/OrganizationSwitcher/OtherOrganizationActions.tsx
@@ -8,10 +8,9 @@ import {
   useCoreUser,
   useOrganizationSwitcherContext,
 } from '../../contexts';
-import { Box, Button, descriptors, Icon, localizationKeys } from '../../customizables';
-import { Action, Actions, OrganizationPreview, PersonalWorkspacePreview } from '../../elements';
-import { useCardState } from '../../elements/contexts';
-import { common, PropsOfComponent } from '../../styledSystem';
+import { Box, descriptors, localizationKeys } from '../../customizables';
+import { Action, Actions, OrganizationPreview, PersonalWorkspacePreview, PreviewButton } from '../../elements';
+import { common } from '../../styledSystem';
 
 type OrganizationActionListProps = {
   onCreateOrganizationClick: React.MouseEventHandler;
@@ -63,10 +62,9 @@ export const OrganizationActionList = (props: OrganizationActionListProps) => {
       >
         {currentOrg && !hidePersonal && (
           <PreviewButton
-            block
-            sx={t => ({
-              height: t.space.$12,
-            })}
+            elementDescriptor={descriptors.organizationSwitcherPreviewButton}
+            icon={SwitchArrows}
+            sx={{ borderRadius: 0 }}
             onClick={onPersonalWorkspaceClick}
           >
             <PersonalWorkspacePreview
@@ -79,11 +77,10 @@ export const OrganizationActionList = (props: OrganizationActionListProps) => {
         )}
         {otherOrgs.map(organization => (
           <PreviewButton
-            block
-            sx={t => ({
-              height: t.space.$12,
-            })}
             key={organization.id}
+            elementDescriptor={descriptors.organizationSwitcherPreviewButton}
+            icon={SwitchArrows}
+            sx={{ borderRadius: 0 }}
             onClick={() => onOrganizationClick(organization)}
           >
             <OrganizationPreview
@@ -96,40 +93,5 @@ export const OrganizationActionList = (props: OrganizationActionListProps) => {
       </Box>
       {createOrganizationButton}
     </Actions>
-  );
-};
-
-const PreviewButton = (props: PropsOfComponent<typeof Button>) => {
-  const { sx, children, ...rest } = props;
-  const card = useCardState();
-
-  return (
-    <Button
-      elementDescriptor={descriptors.organizationSwitcherPreviewButton}
-      variant='ghost'
-      colorScheme='neutral'
-      focusRing={false}
-      hoverAsFocus
-      isDisabled={card.isLoading}
-      sx={[
-        t => ({
-          minHeight: 'unset',
-          borderRadius: 0,
-          justifyContent: 'space-between',
-          padding: `${t.space.$3} ${t.space.$6}`,
-          ':hover > svg': {
-            visibility: 'initial',
-          },
-        }),
-        sx,
-      ]}
-      {...rest}
-    >
-      {children}
-      <Icon
-        icon={SwitchArrows}
-        sx={t => ({ color: t.colors.$blackAlpha500, marginLeft: t.space.$2, visibility: 'hidden' })}
-      />
-    </Button>
   );
 };
