+++ packages/clerk-js/src/ui/elements/index.ts
@@ -53,3 +53,4 @@ export * from './SuccessPage';
 export * from './IconCircle';
 export * from './Popover';
 export * from './Section';
+export * from './PreviewButton';
